requirements.txt
keras==2.9.0
matplotlib==3.5.2
mpld3==0.5.8
numpy==1.21.6
pandas==1.3.5
scikit_learn==1.0.2
streamlit==1.14.0
tensorflow_cpu==2.9.1
yfinance==0.1.85